requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
cryptography==46.0.3
pycparser==2.23
websockets==15.0.1
